Update to .NET 4.5.2 and add a .NET standard 2.0 target
