+++ .github/dependabot.yml
@@ -1,6 +1,6 @@
 version: 2
 updates:
   - package-ecosystem: nuget
-    directory: /lib
+    directory: /
     schedule:
       interval: "daily"
