+++ .gitignore
@@ -2,8 +2,6 @@
 *.suo
 *.user
 *.DotSettings
-lib/packages/*
-lib/D2L.Extensibility.AuthSdk.RestSharp/bin/*
-lib/D2L.Extensibility.AuthSdk.RestSharp/obj/*
-lib/D2L.Extensibility.AuthSdk.RestSharp.Tests/bin/*
-lib/D2L.Extensibility.AuthSdk.RestSharp.Tests/obj/*
+obj/
+bin/
+.vs
