+++ D2L.Extensibility.AuthSdk.Restsharp.Tests/D2L.Extensibility.AuthSdk.Restsharp.Tests.csproj
@@ -1,0 +1,16 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFrameworks>net452;net50</TargetFrameworks>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="D2L.Extensibility.AuthSdk" />
+    <PackageReference Include="Moq" />
+    <PackageReference Include="NUnit" />
+    <PackageReference Include="RestSharp" />
+    <PackageReference Include="NUnit3TestAdapter" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\D2L.Extensibility.AuthSdk.Restsharp\D2L.Extensibility.AuthSdk.Restsharp.csproj" />
+  </ItemGroup>
+</Project>
