+++ D2L.Extensibility.AuthSdk.Restsharp.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31613.86
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "D2L.Extensibility.AuthSdk.Restsharp", "D2L.Extensibility.AuthSdk.Restsharp\D2L.Extensibility.AuthSdk.Restsharp.csproj", "{73526844-BB6D-4BE9-B4CD-852BE1A1B393}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "D2L.Extensibility.AuthSdk.Restsharp.Tests", "D2L.Extensibility.AuthSdk.Restsharp.Tests\D2L.Extensibility.AuthSdk.Restsharp.Tests.csproj", "{A484A34C-FD1B-4EC1-BD11-B838CB577C1F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{73526844-BB6D-4BE9-B4CD-852BE1A1B393}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{73526844-BB6D-4BE9-B4CD-852BE1A1B393}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{73526844-BB6D-4BE9-B4CD-852BE1A1B393}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{73526844-BB6D-4BE9-B4CD-852BE1A1B393}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A484A34C-FD1B-4EC1-BD11-B838CB577C1F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A484A34C-FD1B-4EC1-BD11-B838CB577C1F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A484A34C-FD1B-4EC1-BD11-B838CB577C1F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A484A34C-FD1B-4EC1-BD11-B838CB577C1F}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {719BC18F-911E-405A-82C6-512F867F011E}
+	EndGlobalSection
+EndGlobal
