+++ Directory.Packages.props
@@ -1,0 +1,10 @@
+<Project>
+  <ItemGroup>
+    <PackageVersion Include="D2L.Extensibility.AuthSdk" Version="2.0.0" />
+    <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="16.11.0" />
+    <PackageVersion Include="Moq" Version="4.16.1" />
+    <PackageVersion Include="NUnit" Version="3.13.2" />
+    <PackageVersion Include="NUnit3TestAdapter" Version="4.0.0" />
+    <PackageVersion Include="RestSharp" Version="106.12.0" />
+  </ItemGroup>
+</Project>
