--- lib/D2L.Extensibility.AuthSdk.Restsharp.Tests/packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="D2L.Extensibility.AuthSdk" version="2.0.0" targetFramework="net45" />
-  <package id="Moq" version="4.16.1" targetFramework="net45" />
-  <package id="NUnit" version="3.13.2" targetFramework="net45" />
-  <package id="RestSharp" version="106.12.0" targetFramework="net45" />
-</packages>
