--- lib/D2L.Extensibility.AuthSdk.Restsharp/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="D2L.Extensibility.AuthSdk" version="2.0.0" targetFramework="net40" />
-  <package id="RestSharp" version="106.12.0" targetFramework="net40" />
-</packages>
